requirements.txt
Django
djangorestframework
psycopg2-binary
pillow
drf-yasg
python-decouple
djangorestframework-simplejwt
pyjwt
celery
tzlocal
django-filter
redis
flower